+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This changelog contains the relevant feature additions and bug fixes. To obtain a complete diff between versions you can got to https://github.com/hpcloud/HPCloud-PHP/compare/XXX...XXX where the XXX values are two different tagged versions of the library. For example, https://github.com/hpcloud/HPCloud-PHP/compare/1.0.0-beta6...1.0.0
 
+* X.X.X (xxxx-xx-xx)
+
+  * Removed DBaaS support. The support was for a private beta version with a now deprecated API.
+  * Updated the documentation surrounding CDN and access/secret keys with HP Helion OpenStack and HP Helion Public Cloud.
+  * #14: Fixed bug where transport.debug was ignored by the PHP Stream transport layer.
+
 * 1.2.1 (2013-09-26)
 
   * Disabling HTTP 1.1 keep-alive for the PHP transport layer. PHP doesn't support keep alive.